+++ notebooks/prototyping.py
@@ -1,7 +1,9 @@
+
+
 import marimo
 
 __generated_with = "0.13.2"
-app = marimo.App()
+app = marimo.App(width="medium")
 
 
 @app.cell
@@ -28,13 +30,24 @@ def _():
     warnings.filterwarnings("ignore")
 
     DATA = pl.Path("data")
-    return DATA, gk
+    return DATA, gk, pd
 
 
 @app.cell
-def _(DATA, gk):
+def _(DATA, gk, pd):
     feed = gk.read_feed(DATA / "cairns_gtfs.zip", dist_units="m")
-    feed.shapes
+
+    # Turning a route's shapes into point geometries, should yield an empty route geometry
+    # and should not throw an error
+    rid = feed.routes["route_id"].iat[0]
+    shids = feed.trips.loc[lambda x: x["route_id"] == rid, "shape_id"]
+    f0 = feed.shapes.loc[lambda x: x["shape_id"].isin(shids)].drop_duplicates("shape_id")
+    f1 = feed.shapes.loc[lambda x: ~x["shape_id"].isin(shids)]
+    feed.shapes = pd.concat([f0, f1])
+
+    assert feed.get_routes(as_gdf=True).loc[lambda x: x["route_id"] == rid, "geometry"].iat[0] == None
+
+
     return